requirements.txt
kubernetes
fastapi
uvicorn[standard]
pyjwt[crypto]
gitpython